+++ MoreShipUpgrades/Misc/LGUScanNodeProperties.cs
@@ -48,7 +48,7 @@ namespace MoreShipUpgrades.Misc
         /// <param name="scrapValue">The value of the item in being sold to set on the component</param>
         /// <param name="minRange">The minimum range necessary to be displayed in the player's HUD to set on the component</param>
         /// <param name="maxRange">The maximum range allowed to be displayed in the player's HUD to set on the component</param>
-        public static void ChangeScanNode(ref ScanNodeProperties scanNodeProperties, NodeType nodeType, string header = "LGU Scan Node", string subText = "Used for LGU stuff", int creatureScanID = -1, int scrapValue = 0, int minRange = 2, int maxRange = 7)
+        public static void ChangeScanNode(ref ScanNodeProperties scanNodeProperties, NodeType nodeType, string header = "LGU Scan Node", string subText = "Used for LGU stuff", int creatureScanID = -1, int scrapValue = 0, int minRange = 2, int maxRange = 7, bool requiresLineOfSight = true)
         {
             scanNodeProperties.headerText = header;
             scanNodeProperties.subText = subText;
@@ -57,6 +57,7 @@ namespace MoreShipUpgrades.Misc
             scanNodeProperties.scrapValue = scrapValue;
             scanNodeProperties.minRange = minRange;
             scanNodeProperties.maxRange = maxRange;
+            scanNodeProperties.requiresLineOfSight = requiresLineOfSight;
         }
         /// <summary>
         /// Creates a new gameObject which will contain the scan node and adds it to the provided gameObject
@@ -82,11 +83,11 @@ namespace MoreShipUpgrades.Misc
         /// <param name="creatureScanID">Identifier of the bestiary entry to unlock when scanning this node</param>
         /// <param name="minRange">Minimum range required to display this node</param>
         /// <param name="maxRange">Maximum range allowed to display this node</param>
-        static void AddScanNode(GameObject objectToAddScanNode, NodeType nodeType, string header = "LGU Scan Node", string subText = "Used for LGU stuff", int creatureScanID = -1, int minRange = 2, int maxRange = 7)
+        static void AddScanNode(GameObject objectToAddScanNode, NodeType nodeType, string header = "LGU Scan Node", string subText = "Used for LGU stuff", int creatureScanID = -1, int minRange = 2, int maxRange = 7, bool requiresLineOfSight = true)
         {
             GameObject scanNodeObject = CreateCanvasScanNode(ref objectToAddScanNode);
             ScanNodeProperties scanNodeProperties = scanNodeObject.AddComponent<ScanNodeProperties>();
-            ChangeScanNode(scanNodeProperties: ref scanNodeProperties, nodeType: nodeType, header: header, subText: subText, creatureScanID: creatureScanID, minRange: minRange, maxRange: maxRange);
+            ChangeScanNode(scanNodeProperties: ref scanNodeProperties, nodeType: nodeType, header: header, subText: subText, creatureScanID: creatureScanID, minRange: minRange, maxRange: maxRange, requiresLineOfSight: requiresLineOfSight);
         }
         /// <summary>
         /// Adds a general scan node to the provided gameObject and set its attributes to the provided ones
@@ -97,9 +98,9 @@ namespace MoreShipUpgrades.Misc
         /// <param name="creatureScanID">Identifier of the bestiary entry to unlock when scanning this node</param>
         /// <param name="minRange">Minimum range required to display this node</param>
         /// <param name="maxRange">Maximum range allowed to display this node</param>
-        public static void AddGeneralScanNode(GameObject objectToAddScanNode, string header = "LGU Scan Node", string subText = "Used for LGU stuff", int creatureScanID = -1, int minRange = 2, int maxRange = 7)
+        public static void AddGeneralScanNode(GameObject objectToAddScanNode, string header = "LGU Scan Node", string subText = "Used for LGU stuff", int creatureScanID = -1, int minRange = 2, int maxRange = 7, bool requiresLineOfSight = true)
         {
-            AddScanNode(objectToAddScanNode: objectToAddScanNode, nodeType: NodeType.GENERAL, header: header, subText: subText, creatureScanID: creatureScanID, minRange: minRange, maxRange: maxRange);
+            AddScanNode(objectToAddScanNode: objectToAddScanNode, nodeType: NodeType.GENERAL, header: header, subText: subText, creatureScanID: creatureScanID, minRange: minRange, maxRange: maxRange, requiresLineOfSight: requiresLineOfSight);
         }
         /// <summary>
         /// Destroys the component associated with scan node